chore: rework input suggestions
